[Modified] OC-1087 #comment Made ElementsLinkedList class public #time 30m
